# api/swagger/soccer-manager.swagger.yaml
basePath: /api
definitions:
  ApiResponse:
    properties:
      data: {}
    type: object
...
  UpdatePasswordRequest:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
  UserResponse:
    properties:
      id:
        type: integer
      role:
        type: string
      username:
        type: string
    type: object
  echo.HTTPError:
    properties:
      message: {}
    type: object
host: localhost:80
info:
  contact: {}
  description: Soccer manager api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Soccer Manager
  version: 1.0.0-alpha0
paths:
...
  /v1/users:
    get:
      description: Get a paginated list of users
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/UserResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: List users (ADMIN)
      tags:
      - users
...
  /v1/users/me:
    delete:
      description: Delete the authenticated user's account
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Delete current user
      tags:
      - users
    get:
      description: Get the authenticated user's data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Get current user
      tags:
      - users
  /v1/users/me/change-password:
    put:
      consumes:
      - application/json
      description: Change the authenticated user's password
      parameters:
      - description: Password update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdatePasswordRequest'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Change current user's password
      tags:
      - users
schemes:
- http
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"